+++ src/functions/go.mod
@@ -2,4 +2,7 @@ module github.com/joshua-dev/smartfarmer-backend/src/functions
 
 go 1.14
 
-require cloud.google.com/go/firestore v1.2.0
+require (
+	cloud.google.com/go/firestore v1.2.0
+	google.golang.org/api v0.20.0
+)
