+++ src/functions/insert.go
@@ -1,4 +1,4 @@
-// Package functions defines sensor data model and implements its CRUD operations.
+// Package functions defines sensor data model and implements related operations.
 package functions
 
 import (
@@ -10,14 +10,11 @@ import (
 	"cloud.google.com/go/firestore"
 )
 
-const projectID = "superfarmers"
-
 // Insert stores a sensor data into Firestore.
 // exported to https://asia-northeast1-superfarmers.cloudfunctions.net/Insert
 func Insert(writer http.ResponseWriter, req *http.Request) {
 	ctx := context.Background()
 
-	// create new firestore client and close it when query is done
 	client, err := firestore.NewClient(ctx, projectID)
 	if err != nil {
 		fmt.Fprintf(writer, "firestore.NewClient: %v", err)
@@ -25,25 +22,20 @@ func Insert(writer http.ResponseWriter, req *http.Request) {
 	}
 	defer client.Close()
 
-	// create new sensor data and parse json from request body
-	var data SensorData
-	if err = json.NewDecoder(req.Body).Decode(&data); err != nil {
+	// parse JSON -> sensor data
+	data := new(sensorData)
+	if err = json.NewDecoder(req.Body).Decode(data); err != nil {
 		fmt.Fprintf(writer, "json.Decode: %v", err)
 		return
 	}
 	defer req.Body.Close()
 
+	// update transmission time
 	data.setTime()
 
-	// validates data
-	if err = data.validate(); err != nil {
-		fmt.Fprintf(writer, "validation failed: %v", err)
-	}
-
-	// store data into collection
-	if _, _, err = client.Collection("sensor_data").Add(ctx, data.toMap()); err != nil {
+	// store data into Firestore
+	if _, _, err = client.Collection("sensor_data").Add(ctx, data); err != nil {
 		fmt.Fprintf(writer, "firestore.Add: %v", err)
 		return
 	}
-	fmt.Fprintln(writer, "Successfully stored to Firestore.")
 }
